# perf: eliminate bounds checks in applyMask