--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,3 @@
 module github.com/mccutchen/websocket
 
-go 1.21.0
+go 1.22.0